[Deploy] 추천 도서 리스트 관련 배포입니다.

+++ movie-book/src/main/java/hello/moviebook/Movie/DTO/MovieRatingReq.java
@@ -5,5 +5,5 @@ import lombok.Getter;
 @Getter
 public class MovieRatingReq {
     private Long movieId;
-    private Double rating;
+    private Long rating;
 }
